created file exclusively for cache, renamed previous cache table
