+++ db_cache.py
@@ -1,0 +1,4 @@
+import sqlite3
+
+
+create_table_cache = 'create table if not exists cache ()'
